+++ Reach/src/utopia/reach/container/ReachCanvas2.scala
@@ -1,6 +1,7 @@
 package utopia.reach.container
 
 import utopia.firmament.awt.AwtComponentExtensions._
+import utopia.firmament.awt.AwtEventThread
 import utopia.firmament.component.stack.Stackable
 import utopia.firmament.model.stack.StackSize
 import utopia.flow.collection.mutable.VolatileList
@@ -175,8 +176,8 @@ class ReachCanvas2 protected(contentPointer: Changing[Option[ReachComponentLike]
 	// INITIAL CODE	---------------------------
 	
 	// When bounds get updated, updates the underlying component, also
-	positionPointer.addContinuousListener { e => component.setLocation(e.newValue.toAwtPoint) }
-	sizePointer.addContinuousListener { e => component.setSize(e.newValue.toDimension) }
+	positionPointer.addContinuousListener { e => AwtEventThread.async { component.setLocation(e.newValue.toAwtPoint) } }
+	sizePointer.addContinuousListener { e => AwtEventThread.async { component.setSize(e.newValue.toDimension) } }
 	
 	attachmentPointer.addListener { event =>
 		// When attached to the stack hierarchy, makes sure to update immediate content layout and repaint this component